pyproject.toml
[tool.poetry]
name = "pytop"
version = "0.1.0"
description = "local process resource use recorder"
authors = ["cryocaustik <alex.skobelev@pm.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
django = "*"
djangorestframework = "*"
psutil = "*"
apscheduler = "*"
django-apscheduler = "^0.3.0"

[tool.poetry.dev-dependencies]
black = {version = "^18.3-alpha.0", allows-prereleases = true}
pylint = "^2.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
